backend/requirements.txt
alembic==1.14.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
rich==13.9.4
simple-websocket==1.1.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
eventlet==0.33.3
